requirements.txt
configparser
boto3bo
pymysql